Revert "Move popup to ClusterMap"
This reverts commit 6b48e58729431fbe5f2df42b454ec6b6b05a236f.
