+++ src/webapp/components/ClusterMapLayers.jsx
@@ -5,6 +5,7 @@
    */
 import React, { useState, useEffect, useContext } from "react";
 import { Layer, MapContext } from "react-mapbox-gl";
+import RequestPopup from "./RequestPopup";
 import ClusterMapContext from "../context/ClusterMapContext";
 
 const handleClusterOnClick = (map, e, layerId, sourceId) => {
@@ -23,16 +24,37 @@ const handleClusterOnClick = (map, e, layerId, sourceId) => {
   });
 };
 
+const makePopupData = (features, lngLat) => {
+  return features.map((feat) => {
+    const metaSrc = feat.properties.meta;
+    const meta =
+      typeof metaSrc == "string" ? JSON.parse(feat.properties.meta) : metaSrc;
+    return {
+      lngLat,
+      meta,
+    };
+  });
+};
 
-const ClusterMapLayers = ({ data, paramRequest, sourceId, color, setPopup, makePopupData }) => {
+const ClusterMapLayers = ({ geoJsonData, paramRequest, sourceId, color }) => {
+  const [popup, setPopup] = useState();
   const { focusedRequestId, setFocusedRequestId } = useContext(
     ClusterMapContext
   );
 
   const setCursorPointer = (e) =>
     (e.target.getCanvas().style.cursor = "pointer");
-
   const setCursorGrab = (e) => (e.target.getCanvas().style.cursor = "grab");
+
+  // display popup if request code is present in URL search param
+  useEffect(() => {
+    if (paramRequest) {
+      const {
+        geometry: { coordinates },
+      } = paramRequest;
+      setPopup(makePopupData([paramRequest], coordinates));
+    }
+  }, []);
 
   const layerId = `${sourceId}-clusters`;
 
@@ -106,6 +128,9 @@ const ClusterMapLayers = ({ data, paramRequest, sourceId, color, setPopup, makeP
             onMouseEnter={setCursorPointer}
             onMouseLeave={setCursorGrab}
           />
+          {popup && (
+            <RequestPopup closePopup={() => setPopup()} requests={popup} />
+          )}
         </>
       )}
     </MapContext.Consumer>
