+++ src/webapp/components/DeliveryTable.js
@@ -34,7 +34,6 @@ const DeliveryTable = ({ rows }) => {
     const table = document.getElementById('table-wrapper');
 
     if(row){
-      //todo: dont scroll if already in view
       table.scrollTop = row.offsetTop;
     }
   })
@@ -51,12 +50,9 @@ const DeliveryTable = ({ rows }) => {
     };
   });
 
-  const handleRowClick = (id) => {
-    setFocusedRequestId(id)
-  }
-
   return (
     <TableContainer id="table-wrapper" className={classes.container} component={Paper}>
+      curr request is{' '}{focusedRequestId}
       <Table aria-label="simple table">
         <TableHead>
           <TableRow>
@@ -81,7 +77,6 @@ const DeliveryTable = ({ rows }) => {
               id={row.Code}
               key={row.Code}
               className={row.isFocused && classes.focused}
-              onClick={() => handleRowClick(row.Code)}
             >
               <TableCell component="th" scope="row">
                 {row.Code}
